move actions, refactor makeSelector to functional component
