+++ CarInfo.Frontend/src/components/makeSelector/makeSelector.jsx
@@ -1,35 +1,30 @@
-import { Component } from 'react';
-import * as actions from '../actions.jsx';
+import React, { useEffect, useState } from 'react';
+import { Link } from 'react-router-dom'
+import { getAvailableMakes } from '../../actions/actions.jsx';
 
-class MakeSelector extends Component {
-    constructor(props) {
-        super(props);
-        this.state = { makes: [] }
+function MakeSelector(props) {
+    const [makes, setMakes] = useState(new Array());
+
+    useEffect(() => {
+        getAvailableMakes().then(response => {
+            setMakes(response)
+        })
+    }, [])
+
+    if(makes === []){
+        return(<div>loading...</div>)
     }
 
-    componentDidMount() {
-        actions.getAvailableMakes().then(results => 
-            this.setState(results)
-        );
-    }
-
-    render() { 
-        if(this.state.makes === []){
-            return(<div>loading...</div>)
-        }
-
-        return (
-            <div>
-                {this.state.makes.map(function(item){
-                    console.log(this.state);
-                    return(
-                    <div key={item}>
-                        <a  href={"/models/" + item}>{item}</a>
-                    </div>);
-                }.bind(this))}
-            </div>
-        )
-    }
+    return (
+        <div>
+            {makes.map(item => {
+                return(
+                <div key={item}>
+                    <Link to={"/models/" + item}>{item}</Link>
+                </div>);
+            })}
+        </div>
+    )
 }
 
-export default MakeSelector;
+export default MakeSelector
